+++ src/base.ts
@@ -27,10 +27,14 @@ class ConvoBase {
     return {
       node: this.node,
       apikey: this.apikey,
-      currentVersion: '0.3.17',
+      currentVersion: '0.3.18',
       latestVersion: versionInfo['version'],
       pingResult: pingResult,
     };
+  };
+
+  getWebsocketToken = async (): Promise<any | ErrorType> => {
+    return await fetcher('GET', `${this.node}/getAblyAuth`, this.apikey, {});
   };
 
   ping = async (): Promise<any | ErrorType> => {